--- a/velox/serializers/RowSerializer.h
+++ b/velox/serializers/RowSerializer.h
@@ -406,9 +406,13 @@ class RowDeserializer {
           uncompressedSource, header.uncompressedSize + initialSize);
       while (rowIterator.hasNext()) {
         serializedBuffers.emplace_back(std::move(rowIterator.next()));
-        serializedRows.push_back(std::string_view(
-            serializedBuffers.back()->data(),
-            serializedBuffers.back()->size()));
+        if constexpr (std::is_same_v<SerializeView, std::string_view>) {
+          serializedRows.push_back(std::string_view(
+              serializedBuffers.back()->data(),
+              serializedBuffers.back()->size()));
+        } else {
+          serializedRows.push_back(serializedBuffers.back()->data());
+        }
       }
     }
   }