The old API extracts the single data to std::string_view, which is not efficient, This new approach aligns with CompactRow deserialization, the benchmark shows it accelerates 0% to 5x times depending on the row type.
For backward compatibility, old UnsafeRowDeserializer API calls the new API. The benchmark is not stable, the benchmark result is different depending on the machine status.
```
/mnt/DP_disk1/code/velox/build/velox/row/benchmarks# ./velox_unsafe_row_serialize_benchmark
============================================================================
[...]marks/UnsafeRowSerializeBenchmark.cpp     relative  time/iter   iters/s
============================================================================
unsafe_deserialize_fixedWidth5                            112.69us     8.87K
unsafe_deserialize_fast_fixedWidth5             488.34%    23.08us    43.33K
unsafe_deserialize_fixedWidth10                           124.26us     8.05K
unsafe_deserialize_fast_fixedWidth10            175.44%    70.83us    14.12K
unsafe_deserialize_fixedWidth20                           182.74us     5.47K
unsafe_deserialize_fast_fixedWidth20            99.864%   182.98us     5.46K
decimalsSerialize                                          49.70us    20.12K
decimalsDeserialize                                        70.44us    14.20K
unsafe_deserialize_strings1                                61.31us    16.31K
unsafe_deserialize_fast_strings1                180.84%    33.90us    29.50K
unsafe_deserialize_strings5                               114.08us     8.77K
unsafe_deserialize_fast_strings5                113.19%   100.78us     9.92K
unsafe_deserialize_arrays                                 126.41us     7.91K
unsafe_deserialize_fast_arrays                  373.10%    33.88us    29.51K
unsafe_deserialize_nestedArrays                           350.44us     2.85K
unsafe_deserialize_fast_nestedArrays            259.99%   134.79us     7.42K
unsafe_deserialize_maps                                   536.18us     1.87K
unsafe_deserialize_fast_maps                    500.97%   107.03us     9.34K
unsafe_deserialize_structs                                 87.50us    11.43K
unsafe_deserialize_fast_structs                 147.26%    59.42us    16.83K
```