@jinchengchenghh 2025-01-16
Before change to std::vector<char*>,
```
```
velox/build/velox/row/benchmarks# ./velox_unsafe_row_serialize_benchmark
============================================================================
[...]marks/UnsafeRowSerializeBenchmark.cpp     relative  time/iter   iters/s
============================================================================
unsafe_deserialize_fixedWidth5                             52.83us    18.93K
unsafe_deserialize_fast_fixedWidth5             175.16%    30.16us    33.16K
unsafe_deserialize_fixedWidth10                           273.26us     3.66K
unsafe_deserialize_fast_fixedWidth10            311.01%    87.86us    11.38K
unsafe_deserialize_fixedWidth20                           206.03us     4.85K
unsafe_deserialize_fast_fixedWidth20            205.86%   100.08us     9.99K
unsafe_deserialize_strings1                                54.58us    18.32K
unsafe_deserialize_fast_strings1                156.14%    34.96us    28.61K
unsafe_deserialize_strings5                               204.10us     4.90K
unsafe_deserialize_fast_strings5                159.48%   127.98us     7.81K
unsafe_deserialize_arrays                                 158.01us     6.33K
unsafe_deserialize_fast_arrays                  222.55%    71.00us    14.08K
unsafe_deserialize_nestedArrays                           231.72us     4.32K
unsafe_deserialize_fast_nestedArrays            165.38%   140.11us     7.14K
unsafe_deserialize_maps                                   396.14us     2.52K
unsafe_deserialize_fast_maps                    320.59%   123.56us     8.09K
unsafe_deserialize_structs                                148.35us     6.74K
unsafe_deserialize_fast_structs                 135.53%   109.46us     9.14K
```