@jinchengchenghh 2025-01-16
Run the totally benchmark
```
/mnt/DP_disk1/code/velox/build/velox/row/benchmarks# ./velox_unsafe_row_serialize_benchmark
============================================================================
[...]marks/UnsafeRowSerializeBenchmark.cpp     relative  time/iter   iters/s
============================================================================
unsafe_serialize_fixedWidth5                               50.08us    19.97K
compact_serialize_fixedWidth5                              26.03us    38.42K
container_serialize_fixedWidth5                           150.15us     6.66K
unsafe_deserialize_fixedWidth5                            107.93us     9.27K
unsafe_deserialize_fast_fixedWidth5             248.93%    43.36us    23.06K
compact_deserialize_fixedWidth5                            35.01us    28.56K
container_deserialize_fixedWidth5                         129.72us     7.71K
unsafe_serialize_fixedWidth10                             123.41us     8.10K
compact_serialize_fixedWidth10                             51.83us    19.30K
container_serialize_fixedWidth10                          290.12us     3.45K
unsafe_deserialize_fixedWidth10                           123.29us     8.11K
unsafe_deserialize_fast_fixedWidth10            138.77%    88.84us    11.26K
compact_deserialize_fixedWidth10                           53.39us    18.73K
container_deserialize_fixedWidth10                        205.49us     4.87K
unsafe_serialize_fixedWidth20                             172.48us     5.80K
compact_serialize_fixedWidth20                             77.23us    12.95K
container_serialize_fixedWidth20                          457.88us     2.18K
unsafe_deserialize_fixedWidth20                           500.25us     2.00K
unsafe_deserialize_fast_fixedWidth20            198.19%   252.41us     3.96K
compact_deserialize_fixedWidth20                          104.99us     9.52K
container_deserialize_fixedWidth20                        384.62us     2.60K
decimalsSerialize                                          51.42us    19.45K
decimalsDeserialize                                        73.93us    13.53K
unsafe_serialize_strings1                                  36.94us    27.07K
compact_serialize_strings1                                 27.71us    36.09K
container_serialize_strings1                               90.83us    11.01K
unsafe_deserialize_strings1                                54.76us    18.26K
unsafe_deserialize_fast_strings1                153.17%    35.75us    27.97K
compact_deserialize_strings1                               31.88us    31.36K
container_deserialize_strings1                             84.02us    11.90K
unsafe_serialize_strings5                                 103.82us     9.63K
compact_serialize_strings5                                 78.93us    12.67K
container_serialize_strings5                              413.64us     2.42K
unsafe_deserialize_strings5                               104.09us     9.61K
unsafe_deserialize_fast_strings5                103.84%   100.24us     9.98K
compact_deserialize_strings5                              139.13us     7.19K
container_deserialize_strings5                            236.07us     4.24K
unsafe_serialize_arrays                                    72.52us    13.79K
compact_serialize_arrays                                   58.26us    17.17K
container_serialize_arrays                                145.92us     6.85K
unsafe_deserialize_arrays                                  85.43us    11.70K
unsafe_deserialize_fast_arrays                  235.67%    36.25us    27.59K
compact_deserialize_arrays                                 26.44us    37.82K
container_deserialize_arrays                              137.58us     7.27K
unsafe_serialize_nestedArrays                             329.86us     3.03K
compact_serialize_nestedArrays                            275.41us     3.63K
container_serialize_nestedArrays                          591.96us     1.69K
unsafe_deserialize_nestedArrays                           283.85us     3.52K
unsafe_deserialize_fast_nestedArrays            210.22%   135.03us     7.41K
compact_deserialize_nestedArrays                          107.86us     9.27K
container_deserialize_nestedArrays                        484.08us     2.07K
unsafe_serialize_maps                                     287.61us     3.48K
compact_serialize_maps                                    236.49us     4.23K
container_serialize_maps                                    2.08ms    481.83
unsafe_deserialize_maps                                   425.68us     2.35K
unsafe_deserialize_fast_maps                    290.92%   146.32us     6.83K
compact_deserialize_maps                                  101.10us     9.89K
container_deserialize_maps                                341.35us     2.93K
unsafe_serialize_structs                                   77.17us    12.96K
compact_serialize_structs                                  74.01us    13.51K
container_serialize_structs                               206.36us     4.85K
unsafe_deserialize_structs                                110.84us     9.02K
unsafe_deserialize_fast_structs                 186.71%    59.37us    16.84K
compact_deserialize_structs                                45.18us    22.13K
container_deserialize_structs                             186.85us     5.35K
```