@mbasmanova 2025-03-06
@jinchengchenghh Curious, where / how do you use UnsafeRow serde?