@jinchengchenghh 2025-03-06
When we return the result to Spark, we need the RowVector to row, especially the fallback case, return row to Spark operators and convert row to Vector as velox task input in the next stage